% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{calculate_vpd_and_tair}
\alias{calculate_vpd_and_tair}
\title{Calculate VPD and Tair}
\usage{
calculate_vpd_and_tair(data, rh_col, tair_col)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{rh_col}{Name of the relative humidity column.}

\item{tair_col}{Name of the air temperature column.}
}
\value{
Data frame with calculated VPD and Tair columns added.
}
\description{
Calculates VPD and Tair for the specified data frame.
}
