% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{convert_to_POSIX}
\alias{convert_to_POSIX}
\title{Convert to POSIX time format}
\usage{
convert_to_POSIX(
  data,
  date_format,
  year_col,
  month_col,
  day_col,
  hour_col,
  min_col,
  posix_time_col
)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{date_format}{Character string specifying the date format (e.g., "YMDHM").}

\item{year_col}{Name of the year column.}

\item{month_col}{Name of the month column.}

\item{day_col}{Name of the day column.}

\item{hour_col}{Name of the hour column.}

\item{min_col}{Name of the minute column.}

\item{posix_time_col}{Name of the POSIXct column to be created.}
}
\value{
Data frame with the POSIXct column.
}
\description{
Converts the specified data columns to POSIXct format using the REddyProc package function \code{fConvertTimeToPosix}.
}
