% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{create_cv_folds}
\alias{create_cv_folds}
\title{Create cross-validation folds}
\usage{
create_cv_folds(data, folds = 5, repeats = 2)
}
\arguments{
\item{data}{The dataset}

\item{folds}{Number of folds for cross-validation}

\item{repeats}{Number of repeats for cross-validation}
}
\value{
Cross-validation folds
}
\description{
This function creates cross-validation folds for model training.
}
