% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{create_gap_filled_column}
\alias{create_gap_filled_column}
\title{Create Gap-Filled Column}
\usage{
create_gap_filled_column(data, varname, model)
}
\arguments{
\item{data}{The dataset.}

\item{varname}{The name of the target variable.}

\item{model}{The trained model for predicting missing values.}
}
\value{
The dataset with the new gap-filled column.
}
\description{
This function creates a new column in the dataset where missing values are filled with predictions from a specified model.
}
