% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{create_model_spec}
\alias{create_model_spec}
\title{Create model specification}
\usage{
create_model_spec(
  n_trees = 2000,
  tree_depth = tune(),
  min_n = tune(),
  loss_reduction = tune(),
  sample_size = sample_prop(),
  mtry = tune(),
  learn_rate = tune()
)
}
\arguments{
\item{n_trees}{Number of trees for the model}

\item{tree_depth}{Maximum tree depth}

\item{min_n}{Minimum observations in terminal nodes}

\item{loss_reduction}{Minimum loss reduction to make further splits}

\item{sample_size}{Fraction of the training set to sample}

\item{mtry}{Number of variables to sample at each split}

\item{learn_rate}{Learning rate}
}
\value{
A model specification for XGBoost
}
\description{
This function creates a model specification for the XGBoost model.
}
