% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{create_workflow}
\alias{create_workflow}
\title{Create a workflow}
\usage{
create_workflow(formula, model_spec)
}
\arguments{
\item{formula}{The formula for the model}

\item{model_spec}{The model specification}
}
\value{
A workflow for training the model
}
\description{
This function creates a workflow for training an XGBoost model.
}
