% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{estimate_ustar_threshold_distribution}
\alias{estimate_ustar_threshold_distribution}
\title{Estimate u* threshold distribution for an XGFrictionProcessor object}
\usage{
estimate_ustar_threshold_distribution(
  processor,
  n_sample = 700L,
  probs = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{processor}{XGFrictionProcessor object.}

\item{n_sample}{Number of samples for estimation (default: 700L).}

\item{probs}{Probabilities for estimation (default: c(0.05, 0.5, 0.95)).}
}
\value{
Data frame containing the estimated u* threshold distribution.
}
\description{
This function estimates the u* threshold distribution using the specified processor.
}
