% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{export_results_from_processor}
\alias{export_results_from_processor}
\title{Export results from XGFrictionProcessor object}
\usage{
export_results_from_processor(processor)
}
\arguments{
\item{processor}{XGFrictionProcessor object.}
}
\value{
Data frame containing the exported results.
}
\description{
This function exports the results from the specified processor after gap filling.
}
