% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{filter_ustar_by_mode}
\alias{filter_ustar_by_mode}
\title{Filter u* Threshold Distribution by Aggregation Mode}
\usage{
filter_ustar_by_mode(ustar_distribution, mode = "year")
}
\arguments{
\item{ustar_distribution}{Data frame containing the estimated u* threshold distribution.}

\item{mode}{Character string specifying the aggregation mode ("year", "season", or "single"). Default is "year".}
}
\value{
Data frame filtered based on the specified aggregation mode with selected quantiles (5\%, 50\%, 95\%).
}
\description{
This function filters a data frame containing the estimated u* threshold distribution based on the specified aggregation mode.
}
