% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{gap_fill_mds_variables}
\alias{gap_fill_mds_variables}
\title{Gap Fill MDS Variables}
\usage{
gap_fill_mds_variables(processor, variables, fill_all = FALSE)
}
\arguments{
\item{processor}{XGFriction processing object}

\item{variables}{Vector of variables to gap fill}

\item{fill_all}{Logical indicating whether to fill all gaps}
}
\value{
Updated XGFriction processing object
}
\description{
This function performs gap filling for specified MDS variables using REddyProc.
}
