% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{gap_fill_met}
\alias{gap_fill_met}
\title{Perform gap-filling of meteorological data for XGFrictionProcessor object}
\usage{
gap_fill_met(processor_name, variables, FillAll = FALSE)
}
\arguments{
\item{processor_name}{Name of the XGFrictionProcessor object.}

\item{variables}{Vector of variable names to gap-fill.}

\item{FillAll}{Logical, indicating whether to fill all gaps in the variables or only the missing values between two valid observations.}
}
\value{
XGFrictionProcessor object with gap-filled meteorological data.
}
\description{
This function performs gap-filling of meteorological data using the sMDSGapFill method for an XGFrictionProcessor object.
}
