% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{gap_fill_met_data}
\alias{gap_fill_met_data}
\title{Perform gap-filling of meteorological data for XGFrictionProcessor object}
\usage{
gap_fill_met_data(processor, variables, fill_all = FALSE)
}
\arguments{
\item{processor}{XGFrictionProcessor object to perform gap-filling on.}

\item{variables}{Vector of variable names to gap-fill.}

\item{fill_all}{Logical, indicating whether to fill all gaps in the variables or only the missing values between two valid observations.}
}
\value{
Updated XGFrictionProcessor object with gap-filled meteorological data.
}
\description{
This function performs gap-filling of meteorological data using the \code{sMDSGapFill} method for an XGFrictionProcessor object.
}
