% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{get_ustar_annual}
\alias{get_ustar_annual}
\title{Get Annual u* Threshold Map}
\usage{
get_ustar_annual(filtered_ustar)
}
\arguments{
\item{filtered_ustar}{Data frame filtered based on the specified aggregation mode with selected quantiles (5\%, 50\%, 95\%).}
}
\value{
Data frame containing the annual or seasonal u* threshold map.
}
\description{
This function obtains the annual or seasonal u* threshold map from a filtered data frame.
}
