% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{get_ustar_suffixes}
\alias{get_ustar_suffixes}
\title{Get u* Suffixes from u* Threshold Map}
\usage{
get_ustar_suffixes(uStarTh)
}
\arguments{
\item{uStarTh}{Data frame containing the u* threshold map.}
}
\value{
Character vector containing the column names of the u* threshold map.
}
\description{
This function retrieves the u* suffixes from the u* threshold map.
}
