% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{initialize_user_class_processor}
\alias{initialize_user_class_processor}
\title{Initialize user_class processor}
\usage{
initialize_user_class_processor(
  site_ID,
  data,
  column_names,
  posix_time_column,
  LatDeg,
  LongDeg,
  TimeZoneHour
)
}
\arguments{
\item{site_ID}{Site ID from the AmeriFlux website.}

\item{data}{Data frame containing the eddy data.}

\item{column_names}{Vector with selected column names.}

\item{posix_time_column}{Column name with POSIX time stamp.}

\item{LatDeg}{Latitude of the site.}

\item{LongDeg}{Longitude of the site.}

\item{TimeZoneHour}{Time zone offset in hours.}
}
\value{
Initialized user_class object with location information set.
}
\description{
Initializes the user_class processor object and sets location information.
}
