% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{initialize_xgfriction_processor}
\alias{initialize_xgfriction_processor}
\title{Initialize XGFriction Processor}
\usage{
initialize_xgfriction_processor(site)
}
\arguments{
\item{site}{Site name.}
}
\value{
XGFrictionProcessor object.
}
\description{
This function initializes the XGFriction processor object.
}
