% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{perform_flux_partitioning}
\alias{perform_flux_partitioning}
\title{Perform Flux Partitioning}
\usage{
perform_flux_partitioning(processor, method, params)
}
\arguments{
\item{processor}{The REddyProc object.}

\item{method}{The partitioning method to use. Options: "sGL", "sMR", "sTK".}

\item{params}{List of parameters required for the chosen method.
\itemize{
\item For "sGL" method:
\itemize{
\item useLocaltime: Logical. If TRUE, use local time zone instead of geo-solar time to compute potential radiation.
\item debug.l: Deprecated. Use debug instead.
\item isWarnReplaceColumns: Logical. Set to FALSE to avoid the warning on replacing output columns.
}
\item For "sMR" method:
\itemize{
\item FluxVar: Character. Name of the net ecosystem flux variable.
\item QFFluxVar: Character. Name of the quality-flagged net ecosystem flux variable.
\item QFFluxValue: Numeric. Quality flag value to indicate bad flux values.
\item TempVar: Character. Name of the temperature variable.
\item QFTempVar: Character. Name of the quality-flagged temperature variable.
\item QFTempValue: Numeric. Quality flag value to indicate bad temperature values.
\item RadVar: Character. Name of the radiation variable.
\item TRef: Numeric. Reference temperature.
\item Suffix.s: Character. Suffix for the partitioning results.
}
\item For "sTK" method:
\itemize{
\item controlGLPart: Further default parameters, such as suffix.
}
}}
}
\description{
This function performs flux partitioning using the specified method and parameters.
}
