% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{perform_gap_filling_with_ustar}
\alias{perform_gap_filling_with_ustar}
\title{Perform Gap Filling using u* Threshold Distribution}
\usage{
perform_gap_filling_with_ustar(
  processor,
  variable,
  ustar_distribution,
  fill_all = TRUE
)
}
\arguments{
\item{processor}{An instance of the sEddyProc class or similar processor object containing the user's data.}

\item{variable}{Character string specifying the variable to gap fill (e.g., "NEE").}

\item{ustar_distribution}{Data frame containing the estimated u* threshold distribution. It should include the quantiles (5\%, 50\%, 95\%) for the specified aggregation mode (e.g., "year").}

\item{fill_all}{Logical indicating whether to fill all gaps (default: TRUE). If \code{TRUE}, all gaps in the specified variable will be filled using the u* threshold distribution. If \code{FALSE}, only gaps matching specific conditions will be filled.}
}
\value{
Updated processor object after gap filling. The processor object will have the specified variable gap-filled using the u* threshold distribution.
}
\description{
This function performs gap filling for a specified variable using the u* threshold distribution.
}
