% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{perform_iqr_filtering}
\alias{perform_iqr_filtering}
\title{Perform IQR Filtering}
\usage{
perform_iqr_filtering(
  xgfriction_proc,
  dataframe,
  variables,
  threshold_multiplier = 6
)
}
\arguments{
\item{dataframe}{Data frame containing the variables to filter}

\item{variables}{Vector of variables to filter}

\item{threshold_multiplier}{Multiplier for IQR threshold}

\item{processor}{XGFriction processing object}
}
\value{
Updated XGFriction processing object
}
\description{
This function performs IQR filtering on specified variables.
}
