% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{perform_ustar_gap_fill}
\alias{perform_ustar_gap_fill}
\title{Perform Ustar Threshold Distribution Estimation and Gap Filling}
\usage{
perform_ustar_gap_fill(
  processor,
  dataframe,
  variable,
  nSample = 1000L,
  probs = c(0.05, 0.5, 0.95),
  fill_all = TRUE
)
}
\arguments{
\item{processor}{XGFriction processing object}

\item{dataframe}{Data frame containing the variable to perform gap filling on}

\item{variable}{Variable to perform gap filling on}

\item{nSample}{Number of samples for estimation}

\item{probs}{Probabilities for estimation}

\item{fill_all}{Logical indicating whether to fill all gaps}
}
\value{
Updated XGFriction processing object
}
\description{
This function estimates the u* threshold distribution and performs gap filling.
}
