% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{plot_model_metrics}
\alias{plot_model_metrics}
\title{Plot Model Metrics}
\usage{
plot_model_metrics(metrics, metric_name)
}
\arguments{
\item{metrics}{The metrics data frame.}

\item{metric_name}{The name of the metric to plot.}
}
\value{
A plot showing the specified metric.
}
\description{
This function plots the metrics of a trained XGBoost model.
}
