% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{plot_variables_of_importance}
\alias{plot_variables_of_importance}
\title{Plot Variables of Importance}
\usage{
plot_variables_of_importance(model, n_top = 10)
}
\arguments{
\item{model}{The trained XGBoost model.}

\item{n_top}{Number of top variables to plot.}
}
\value{
A plot showing the variables of importance.
}
\description{
This function plots the variables of importance of a trained XGBoost model.
}
