% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{predict_missing_data}
\alias{predict_missing_data}
\title{Use trained XGBoost model to predict missing data}
\usage{
predict_missing_data(model, data)
}
\arguments{
\item{model}{The trained XGBoost model.}

\item{data}{The dataset with missing values.}
}
\value{
The dataset with missing values filled using the XGBoost model.
}
\description{
This function uses a trained XGBoost model to predict missing data in a dataset.
}
