% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for model training}
\usage{
prepare_data(data)
}
\arguments{
\item{data}{The dataset to be split}
}
\value{
A list containing the training and testing sets
}
\description{
This function splits the data into training and testing sets.
}
