% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{replace_variables_in_REddyProc}
\alias{replace_variables_in_REddyProc}
\title{Replace Variables in REddyProc Object}
\usage{
replace_variables_in_REddyProc(processor, data, varnames)
}
\arguments{
\item{processor}{The REddyProc object.}

\item{data}{The dataset containing columns to replace variables with.}

\item{varnames}{Vector of variable names to replace in the REddyProc object.}
}
\description{
This function replaces specified variables in a REddyProc object with columns from a dataset.
}
