% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{select_best_model}
\alias{select_best_model}
\title{Select Best Model}
\usage{
select_best_model(results, metric, higher_is_better = TRUE)
}
\arguments{
\item{results}{A data frame containing the results of model training (e.g., cross-validation results).}

\item{metric}{The evaluation metric used for selecting the best model (e.g., "rmse", "rsq").}

\item{higher_is_better}{Logical indicating whether a higher value of the evaluation metric is better.}
}
\value{
The best model based on the specified evaluation metric.
}
\description{
This function selects the best model based on specified evaluation metrics.
}
