% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{subtract_15_minutes_from_datetime}
\alias{subtract_15_minutes_from_datetime}
\title{Subtract 15 minutes from DateTime column}
\usage{
subtract_15_minutes_from_datetime(data, posix_time_col)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{posix_time_col}{Name of the POSIXct column (DateTime) to adjust.}
}
\value{
Data frame with adjusted DateTime column.
}
\description{
Subtracts 15 minutes (900 seconds) from the DateTime column of a data frame.
}
