% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{train_xgboost_model}
\alias{train_xgboost_model}
\title{Train XGBoost model with custom grid}
\usage{
train_xgboost_model(
  data,
  formula,
  grid,
  folds = 5,
  repeats = 2,
  n_trees = 2000,
  tree_depth = tune(),
  min_n = tune(),
  loss_reduction = tune(),
  sample_size = sample_prop(),
  mtry = tune(),
  learn_rate = tune(),
  seed_count = 1,
  seed_for_data = 60723
)
}
\arguments{
\item{data}{The dataset}

\item{formula}{The formula for the model}

\item{grid}{Custom grid for hyperparameter tuning}

\item{folds}{Number of folds for cross-validation}

\item{repeats}{Number of repeats for cross-validation}

\item{n_trees}{Number of trees for the model}

\item{tree_depth}{Maximum tree depth}

\item{min_n}{Minimum observations in terminal nodes}

\item{loss_reduction}{Minimum loss reduction to make further splits}

\item{sample_size}{Fraction of the training set to sample}

\item{mtry}{Number of variables to sample at each split}

\item{learn_rate}{Learning rate}

\item{seed_count}{Number of seed values to use}

\item{seed_for_data}{Seed value for data preparation}
}
\value{
Trained XGBoost model and testing data
}
\description{
This function trains an XGBoost model using the provided data and a custom grid for hyperparameter tuning.
}
