% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\name{XgfRiction_initialize}
\alias{XgfRiction_initialize}
\title{Initialize sEddyProc object, adjust timestamps, and set location information}
\usage{
XgfRiction_initialize(
  site_ID,
  data,
  column_names,
  posix_time_column,
  LatDeg,
  LongDeg,
  TimeZoneHour
)
}
\arguments{
\item{site_ID}{Site ID from the AmeriFlux website.}

\item{data}{Data frame containing the eddy data.}

\item{column_names}{Vector with selected column names.}

\item{posix_time_column}{Column name with POSIX time stamp.}

\item{LatDeg}{Latitude of the site.}

\item{LongDeg}{Longitude of the site.}

\item{TimeZoneHour}{Time zone offset in hours.}
}
\value{
Initialized sEddyProc object.
}
\description{
This function initializes the sEddyProc class, adjusts timestamps if necessary, and sets location information.
}
