% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XgfRiction.R
\docType{class}
\name{XGFrictionProcessor-class}
\alias{XGFrictionProcessor-class}
\alias{XGFrictionProcessor}
\title{This class represents a processor for XGFriction data.
XGFrictionProcessor Class}
\description{
This class represents a processor for XGFriction data.
}
\section{Fields}{

\describe{
\item{\code{site}}{The name of the site associated with the processor.}
}}


